test: split native runtime policy encoding tests
