+++ packages/jazz-tools/src/runtime/native-runtime/runtime.test.ts
@@ -1,5 +1,4 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { readFileSync, writeFileSync } from "node:fs";
 import { performance } from "node:perf_hooks";
 import type { ColumnDescriptor, NativeRowDelta, WasmSchema } from "../../drivers/types.js";
 import {
@@ -25,12 +24,15 @@ import {
 } from "./native-runtime-adapter.js";
 import { encodeSchema } from "./schema-codec.js";
 import {
+  readSchemaColumnLargeValues,
+  readSchemaSelectPolicyBranches,
+  readSchemaTableMetadata,
+} from "./native-runtime-policy.test-support.js";
+import {
   applySubscriptionDelta,
   decodeNativeDelta,
   SubscriptionManager,
 } from "../subscription-manager.js";
-import { definePermissions } from "../../permissions/index.js";
-import { mergePermissionsIntoWasmSchema } from "../../schema-permissions.js";
 import { setNamedRowValuesEnumerable } from "./row-values-transport.js";
 import { encodeNativeNullValue, storageColumnValueType } from "./native-row-codec.js";
 import { type BatchId, type WriteReceipt } from "../client.js";
@@ -4012,981 +4014,6 @@ describe("NativeRuntimeAdapter server transport", () => {
       offset: 5,
     });
   });
-
-  it("serializes OR policy branches with Exists source id correlation", () => {
-    const policy = readSchemaSelectPolicyBranches(
-      encodeSchema({
-        documents: {
-          columns: [
-            { name: "visibility", column_type: { type: "Text" }, nullable: false },
-            { name: "title", column_type: { type: "Text" }, nullable: false },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "Or",
-                exprs: [
-                  {
-                    type: "Cmp",
-                    column: "visibility",
-                    op: "Eq",
-                    value: { type: "Literal", value: { type: "Text", value: "public" } },
-                  },
-                  {
-                    type: "Exists",
-                    table: "document_members",
-                    condition: {
-                      type: "And",
-                      exprs: [
-                        {
-                          type: "Cmp",
-                          column: "document_id",
-                          op: "Eq",
-                          value: {
-                            type: "SessionRef",
-                            path: ["__jazz_outer_row", "id"],
-                          },
-                        },
-                        {
-                          type: "Cmp",
-                          column: "role",
-                          op: "Eq",
-                          value: { type: "Literal", value: { type: "Text", value: "reader" } },
-                        },
-                      ],
-                    },
-                  },
-                ],
-              },
-            },
-          },
-        },
-        document_members: {
-          columns: [
-            { name: "document_id", column_type: { type: "Uuid" }, nullable: false },
-            { name: "role", column_type: { type: "Text" }, nullable: false },
-          ],
-        },
-      }),
-      "documents",
-    );
-
-    expect(policy).toEqual({
-      table: "documents",
-      filters: [{ tag: 1, children: [] }],
-      joins: [],
-      branches: [
-        {
-          filters: [
-            {
-              tag: 3,
-              column: "visibility",
-              operand: { tag: 3, literalTag: 6, value: "public" },
-            },
-          ],
-          joins: [],
-        },
-        {
-          filters: [],
-          joins: [
-            {
-              table: "document_members",
-              onColumn: "document_id",
-              targetTag: 0,
-              sourceColumn: "id",
-              sourceLookup: undefined,
-              filters: [
-                {
-                  tag: 3,
-                  column: "role",
-                  operand: { tag: 3, literalTag: 6, value: "reader" },
-                },
-              ],
-              nestedJoins: [],
-            },
-          ],
-        },
-      ],
-    });
-  });
-
-  it("serializes same-table gather seeds as seeded reachable policies", () => {
-    const reachables = readSchemaSelectPolicyReachables(
-      encodeSchema({
-        resources: {
-          columns: [{ name: "label", column_type: { type: "Text" }, nullable: false }],
-          policies: {
-            select: {
-              using: {
-                type: "ExistsRel",
-                rel: {
-                  Filter: {
-                    input: {
-                      Join: {
-                        left: {
-                          Gather: {
-                            seed: {
-                              Filter: {
-                                input: { TableScan: { table: "teams" } },
-                                predicate: {
-                                  Cmp: {
-                                    left: { scope: "teams", column: "identity_key" },
-                                    op: "Eq",
-                                    right: { SessionRef: ["userId"] },
-                                  },
-                                },
-                              },
-                            },
-                            step: {
-                              Project: {
-                                input: {
-                                  Join: {
-                                    left: {
-                                      Filter: {
-                                        input: { TableScan: { table: "team_entries" } },
-                                        predicate: {
-                                          And: [
-                                            {
-                                              Cmp: {
-                                                left: {
-                                                  scope: "team_entries",
-                                                  column: "member_id",
-                                                },
-                                                op: "Eq",
-                                                right: { RowId: "Frontier" },
-                                              },
-                                            },
-                                            {
-                                              Cmp: {
-                                                left: {
-                                                  scope: "team_entries",
-                                                  column: "administrator",
-                                                },
-                                                op: "Eq",
-                                                right: {
-                                                  Literal: { type: "Boolean", value: false },
-                                                },
-                                              },
-                                            },
-                                          ],
-                                        },
-                                      },
-                                    },
-                                    right: {
-                                      TableScan: {
-                                        table: "teams",
-                                        alias: "__recursive_hop_0",
-                                      },
-                                    },
-                                    on: [
-                                      {
-                                        left: { scope: "team_entries", column: "target_id" },
-                                        right: { scope: "__recursive_hop_0", column: "id" },
-                                      },
-                                    ],
-                                    join_kind: "Inner",
-                                  },
-                                },
-                                columns: [],
-                              },
-                            },
-                            frontier_key: { RowId: "Current" },
-                            bound: { MaxDepth: 8 },
-                            dedupe_key: [{ RowId: "Current" }],
-                          },
-                        },
-                        right: { TableScan: { table: "resource_access", alias: "access" } },
-                        on: [
-                          {
-                            left: { column: "id" },
-                            right: { scope: "access", column: "team" },
-                          },
-                        ],
-                        join_kind: "Inner",
-                      },
-                    },
-                    predicate: {
-                      And: [
-                        {
-                          Cmp: {
-                            left: { scope: "access", column: "resource" },
-                            op: "Eq",
-                            right: { RowId: "Outer" },
-                          },
-                        },
-                        {
-                          Cmp: {
-                            left: { scope: "access", column: "administrator" },
-                            op: "Eq",
-                            right: { Literal: { type: "Boolean", value: false } },
-                          },
-                        },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-        teams: {
-          columns: [{ name: "identity_key", column_type: { type: "Uuid" }, nullable: false }],
-        },
-        team_entries: {
-          columns: [
-            {
-              name: "member_id",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "teams",
-            },
-            {
-              name: "target_id",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "teams",
-            },
-            { name: "administrator", column_type: { type: "Boolean" }, nullable: false },
-          ],
-        },
-        resource_access: {
-          columns: [
-            {
-              name: "resource",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "resources",
-            },
-            {
-              name: "team",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "teams",
-            },
-            { name: "administrator", column_type: { type: "Boolean" }, nullable: false },
-          ],
-        },
-      }),
-      "resources",
-    );
-
-    expect(reachables).toHaveLength(1);
-    expect(reachables[0]).toMatchObject({
-      accessTable: "resource_access",
-      accessRowColumn: "resource",
-      accessTeamColumn: "team",
-      accessTeamTargetTag: 0,
-      edgeTable: "team_entries",
-      edgeMemberColumn: "member_id",
-      edgeParentColumn: "target_id",
-      maxDepth: 8,
-      seed: {
-        table: "teams",
-        userColumn: "identity_key",
-        userClaim: "user_id",
-        teamColumn: "id",
-        filters: [],
-      },
-    });
-    expect(reachables[0]!.accessFilters).toEqual([
-      {
-        tag: 3,
-        column: "administrator",
-        operand: { tag: 3, literalTag: 5, value: false },
-      },
-    ]);
-    expect(reachables[0]!.edgeFilters).toEqual([
-      {
-        tag: 3,
-        column: "administrator",
-        operand: { tag: 3, literalTag: 5, value: false },
-      },
-    ]);
-  });
-
-  it("serializes reachable_via seeded_by TS policies as the Rust reachable atom", () => {
-    const baseSchema: WasmSchema = {
-      resources: {
-        columns: [{ name: "label", column_type: { type: "Text" }, nullable: false }],
-      },
-      teams: {
-        columns: [{ name: "identity_key", column_type: { type: "Text" }, nullable: false }],
-      },
-      team_team_edges: {
-        columns: [
-          {
-            name: "child_team",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "teams",
-          },
-          {
-            name: "parent_team",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "teams",
-          },
-        ],
-      },
-      resource_access_edges: {
-        columns: [
-          {
-            name: "resource",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "resources",
-          },
-          {
-            name: "team",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "teams",
-          },
-          { name: "grant_role", column_type: { type: "Text" }, nullable: false },
-        ],
-      },
-    };
-    const app = {
-      wasmSchema: baseSchema,
-      resources: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      teams: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      team_team_edges: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      resource_access_edges: { _rowType: {} as never, where: (_input: unknown) => undefined },
-    };
-    const permissions = definePermissions(app, ({ policy, session }) => {
-      policy.resources.allowRead.where(
-        policy.exists(
-          policy.resources
-            .reachable_via_with_access_filters(
-              "resource_access_edges",
-              "resource",
-              "team",
-              session.sub,
-              { grant_role: "viewer" },
-              "team_team_edges",
-              "child_team",
-              "parent_team",
-            )
-            .seeded_by("teams", "identity_key", "sub", "id"),
-        ),
-      );
-    });
-    const reachables = readSchemaSelectPolicyReachables(
-      encodeSchema(mergePermissionsIntoWasmSchema(baseSchema, permissions)),
-      "resources",
-    );
-
-    expect(reachables).toHaveLength(1);
-    expect(reachables[0]).toMatchObject({
-      accessTable: "resource_access_edges",
-      accessRowColumn: "resource",
-      accessTeamColumn: "team",
-      accessTeamTargetTag: 0,
-      edgeTable: "team_team_edges",
-      edgeMemberColumn: "child_team",
-      edgeParentColumn: "parent_team",
-      maxDepth: 8,
-      seed: {
-        table: "teams",
-        userColumn: "identity_key",
-        userClaim: "sub",
-        teamColumn: "id",
-        filters: [],
-      },
-    });
-    expect(reachables[0]!.accessFilters).toEqual([
-      {
-        tag: 3,
-        column: "grant_role",
-        operand: { tag: 3, literalTag: 6, value: "viewer" },
-      },
-    ]);
-  });
-
-  it("serializes allowedTo.read as a native inherits policy atom", () => {
-    const baseSchema: WasmSchema = {
-      resources: {
-        columns: [{ name: "label", column_type: { type: "Text" }, nullable: false }],
-      },
-      data_entries: {
-        columns: [
-          {
-            name: "resource",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "resources",
-          },
-          { name: "label", column_type: { type: "Text" }, nullable: false },
-        ],
-      },
-    };
-    const app = {
-      wasmSchema: baseSchema,
-      resources: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      data_entries: { _rowType: {} as never, where: (_input: unknown) => undefined },
-    };
-    const permissions = definePermissions(app, ({ policy, allowedTo }) => {
-      policy.resources.allowRead.where({ label: "visible" });
-      policy.data_entries.allowRead.where(allowedTo.read("resource"));
-    });
-
-    const policy = readSchemaSelectPolicyInherits(
-      encodeSchema(mergePermissionsIntoWasmSchema(baseSchema, permissions)),
-      "data_entries",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "resource" }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes allowedTo insert, update, and delete as native inherits policy atoms", () => {
-    const baseSchema: WasmSchema = {
-      resources: {
-        columns: [{ name: "label", column_type: { type: "Text" }, nullable: false }],
-      },
-      data_entries: {
-        columns: [
-          {
-            name: "resource",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "resources",
-          },
-          { name: "label", column_type: { type: "Text" }, nullable: false },
-        ],
-      },
-    };
-    const app = {
-      wasmSchema: baseSchema,
-      resources: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      data_entries: { _rowType: {} as never, where: (_input: unknown) => undefined },
-    };
-    const permissions = definePermissions(app, ({ policy, allowedTo }) => {
-      policy.data_entries.allowInsert.where(allowedTo.insert("resource"));
-      policy.data_entries.allowUpdate
-        .whereOld(allowedTo.update("resource"))
-        .whereNew(allowedTo.update("resource"));
-      policy.data_entries.allowDelete.where(allowedTo.delete("resource"));
-    });
-
-    const encoded = encodeSchema(mergePermissionsIntoWasmSchema(baseSchema, permissions));
-
-    expect(readSchemaPolicyInherits(encoded, "data_entries", "insert")).toEqual({
-      inherits: [{ parentColumn: "resource" }],
-      joinCount: 0,
-    });
-    expect(readSchemaPolicyInherits(encoded, "data_entries", "updateUsing")).toEqual({
-      inherits: [{ parentColumn: "resource" }],
-      joinCount: 0,
-    });
-    expect(readSchemaPolicyInherits(encoded, "data_entries", "updateCheck")).toEqual({
-      inherits: [{ parentColumn: "resource" }],
-      joinCount: 0,
-    });
-    expect(readSchemaPolicyInherits(encoded, "data_entries", "delete")).toEqual({
-      inherits: [{ parentColumn: "resource" }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes authored inherited policies byte-identically to native schema atoms", () => {
-    const baseSchema: WasmSchema = {
-      resources: {
-        columns: [{ name: "label", column_type: { type: "Text" }, nullable: false }],
-        policies: {
-          select: { using: { type: "True" } },
-          insert: { with_check: { type: "True" } },
-          update: { using: { type: "True" }, with_check: { type: "True" } },
-          delete: { using: { type: "True" } },
-        },
-      },
-      data_entries: {
-        columns: [
-          {
-            name: "resource",
-            column_type: { type: "Uuid" },
-            nullable: false,
-            references: "resources",
-          },
-          { name: "label", column_type: { type: "Text" }, nullable: false },
-        ],
-      },
-    };
-    const app = {
-      wasmSchema: baseSchema,
-      resources: { _rowType: {} as never, where: (_input: unknown) => undefined },
-      data_entries: { _rowType: {} as never, where: (_input: unknown) => undefined },
-    };
-    const permissions = definePermissions(app, ({ policy, allowedTo }) => {
-      policy.data_entries.allowRead.where(allowedTo.read("resource"));
-      policy.data_entries.allowInsert.where(allowedTo.insert("resource"));
-      policy.data_entries.allowUpdate
-        .whereOld(allowedTo.update("resource"))
-        .whereNew(allowedTo.update("resource"));
-      policy.data_entries.allowDelete.where(allowedTo.delete("resource"));
-    });
-    const nativeSchema: WasmSchema = {
-      ...baseSchema,
-      data_entries: {
-        ...baseSchema.data_entries,
-        policies: {
-          select: {
-            using: { type: "Inherits", operation: "Select", via_column: "resource" },
-          },
-          insert: {
-            with_check: { type: "Inherits", operation: "Insert", via_column: "resource" },
-          },
-          update: {
-            using: { type: "Inherits", operation: "Update", via_column: "resource" },
-            with_check: { type: "Inherits", operation: "Update", via_column: "resource" },
-          },
-          delete: {
-            using: { type: "Inherits", operation: "Delete", via_column: "resource" },
-          },
-        },
-      },
-    };
-
-    expect(encodeSchema(mergePermissionsIntoWasmSchema(baseSchema, permissions))).toEqual(
-      encodeSchema(nativeSchema),
-    );
-  });
-
-  it("encodes the policy graph perf fixture byte-stably", () => {
-    const fixtureDir = new URL("../../testing/fixtures/policy-graph-perf/", import.meta.url);
-    const source = JSON.parse(readFileSync(new URL("schema-source.json", fixtureDir), "utf8")) as {
-      mergedSchema: WasmSchema;
-    };
-    const expectedBytes = new Uint8Array(readFileSync(new URL("schema.native.bin", fixtureDir)));
-    const encoded = encodeSchema(source.mergedSchema);
-
-    if (process.env.JAZZ_UPDATE_POLICY_GRAPH_PERF_NATIVE_SCHEMA) {
-      writeFileSync(new URL("schema.native.bin", fixtureDir), encoded);
-      writeFileSync(
-        new URL("schema.native.hex", fixtureDir),
-        `${Buffer.from(encoded).toString("hex")}\n`,
-      );
-      return;
-    }
-
-    expect(encoded).toEqual(expectedBytes);
-  });
-
-  it("rejects ExistsRel Gather policies without a concrete MaxDepth bound", () => {
-    expect(() =>
-      encodeSchema({
-        teams: {
-          columns: [
-            {
-              name: "parent_id",
-              column_type: { type: "Uuid" },
-              nullable: true,
-              references: "teams",
-            },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "ExistsRel",
-                rel: {
-                  Gather: {
-                    seed: {
-                      Project: {
-                        input: {
-                          Filter: {
-                            input: {
-                              Join: {
-                                left: { TableScan: { table: "teams", alias: "edge" } },
-                                right: { TableScan: { table: "teams", alias: "seed" } },
-                                on: [
-                                  {
-                                    left: { scope: "edge", column: "parent_id" },
-                                    right: { scope: "seed", column: "id" },
-                                  },
-                                ],
-                                join_kind: "Inner",
-                              },
-                            },
-                            predicate: {
-                              Cmp: {
-                                left: { scope: "seed", column: "parent_id" },
-                                op: "Eq",
-                                right: { SessionRef: ["teamId"] },
-                              },
-                            },
-                          },
-                        },
-                        columns: [],
-                      },
-                    },
-                    step: {
-                      Project: {
-                        input: {
-                          Join: {
-                            left: {
-                              Filter: {
-                                input: { TableScan: { table: "teams", alias: "edge" } },
-                                predicate: {
-                                  Cmp: {
-                                    left: { scope: "edge", column: "id" },
-                                    op: "Eq",
-                                    right: { RowId: "Frontier" },
-                                  },
-                                },
-                              },
-                            },
-                            right: { TableScan: { table: "teams", alias: "next" } },
-                            on: [
-                              {
-                                left: { scope: "edge", column: "parent_id" },
-                                right: { scope: "next", column: "id" },
-                              },
-                            ],
-                            join_kind: "Inner",
-                          },
-                        },
-                        columns: [],
-                      },
-                    },
-                    frontier_key: { RowId: "Frontier" },
-                    bound: "Fixpoint",
-                    dedupe_key: [{ RowId: "Current" }],
-                  },
-                },
-              } as never,
-            },
-          },
-        },
-      }),
-    ).toThrow("MaxDepth");
-  });
-
-  it("serializes InheritsReferencing without a source operation policy as fail-closed", () => {
-    const policy = readSchemaSelectPolicyBranches(
-      encodeSchema({
-        projects: {
-          columns: [{ name: "name", column_type: { type: "Text" }, nullable: false }],
-          policies: {
-            select: {
-              using: {
-                type: "InheritsReferencing",
-                operation: "Select",
-                source_table: "todos",
-                via_column: "project_id",
-              },
-            },
-          },
-        },
-        todos: {
-          columns: [
-            {
-              name: "project_id",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "projects",
-            },
-            { name: "title", column_type: { type: "Text" }, nullable: false },
-          ],
-        },
-      }),
-      "projects",
-    );
-
-    expect(policy).toEqual({
-      table: "projects",
-      filters: [],
-      joins: [
-        {
-          table: "todos",
-          onColumn: "project_id",
-          targetTag: 0,
-          sourceColumn: undefined,
-          sourceLookup: undefined,
-          filters: [{ tag: 1, children: [] }],
-          nestedJoins: [],
-        },
-      ],
-      branches: [],
-    });
-  });
-
-  it("serializes direct Inherits delete as a native inherits policy atom", () => {
-    const policy = readSchemaPolicyInherits(
-      encodeSchema({
-        messages: {
-          columns: [{ name: "room_id", column_type: { type: "Uuid" }, nullable: false }],
-          policies: {
-            delete: {
-              using: {
-                type: "Cmp",
-                column: "room_id",
-                op: "Eq",
-                value: { type: "SessionRef", path: ["roomId"] },
-              },
-            },
-          },
-        },
-        reactions: {
-          columns: [
-            {
-              name: "message_id",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "messages",
-            },
-          ],
-          policies: {
-            delete: {
-              using: {
-                type: "Inherits",
-                operation: "Delete",
-                via_column: "message_id",
-              },
-            },
-          },
-        },
-      }),
-      "reactions",
-      "delete",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "message_id" }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes a bounded Inherits depth as a native inherits policy atom", () => {
-    const policy = readSchemaSelectPolicyInherits(
-      encodeSchema({
-        folders: {
-          columns: [
-            {
-              name: "parent_id",
-              column_type: { type: "Uuid" },
-              nullable: true,
-              references: "folders",
-            },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "Inherits",
-                operation: "Select",
-                via_column: "parent_id",
-                max_depth: 3,
-              },
-            },
-          },
-        },
-      }),
-      "folders",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "parent_id", maxDepth: 3 }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes direct Inherits through parent exists joins with source lookup", () => {
-    const policy = readSchemaSelectPolicyInherits(
-      encodeSchema({
-        chats: {
-          columns: [{ name: "isPublic", column_type: { type: "Boolean" }, nullable: false }],
-          policies: {
-            select: {
-              using: {
-                type: "Exists",
-                table: "chatMembers",
-                condition: {
-                  type: "And",
-                  exprs: [
-                    {
-                      type: "Cmp",
-                      column: "chatId",
-                      op: "Eq",
-                      value: { type: "OuterRowRef", column: "id" } as never,
-                    },
-                    {
-                      type: "Cmp",
-                      column: "userId",
-                      op: "Eq",
-                      value: { type: "SessionRef", path: ["user_id"] },
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        },
-        chatMembers: {
-          columns: [
-            {
-              name: "chatId",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "chats",
-            },
-            { name: "userId", column_type: { type: "Text" }, nullable: false },
-          ],
-        },
-        canvases: {
-          columns: [
-            {
-              name: "chatId",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "chats",
-            },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "Inherits",
-                operation: "Select",
-                via_column: "chatId",
-              },
-            },
-          },
-        },
-      }),
-      "canvases",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "chatId" }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes nested Inherits through composed source lookups", () => {
-    const policy = readSchemaSelectPolicyInherits(
-      encodeSchema({
-        chats: {
-          columns: [{ name: "isPublic", column_type: { type: "Boolean" }, nullable: false }],
-          policies: {
-            select: {
-              using: {
-                type: "Exists",
-                table: "chatMembers",
-                condition: {
-                  type: "And",
-                  exprs: [
-                    {
-                      type: "Cmp",
-                      column: "chatId",
-                      op: "Eq",
-                      value: { type: "OuterRowRef", column: "id" } as never,
-                    },
-                    {
-                      type: "Cmp",
-                      column: "userId",
-                      op: "Eq",
-                      value: { type: "SessionRef", path: ["user_id"] },
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        },
-        chatMembers: {
-          columns: [
-            {
-              name: "chatId",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "chats",
-            },
-            { name: "userId", column_type: { type: "Text" }, nullable: false },
-          ],
-        },
-        canvases: {
-          columns: [
-            {
-              name: "chatId",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "chats",
-            },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "Inherits",
-                operation: "Select",
-                via_column: "chatId",
-              },
-            },
-          },
-        },
-        strokes: {
-          columns: [
-            {
-              name: "canvasId",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "canvases",
-            },
-          ],
-          policies: {
-            select: {
-              using: {
-                type: "Inherits",
-                operation: "Select",
-                via_column: "canvasId",
-              },
-            },
-          },
-        },
-      }),
-      "strokes",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "canvasId" }],
-      joinCount: 0,
-    });
-  });
-
-  it("serializes direct Inherits without a parent operation policy as a native inherits policy atom", () => {
-    const policy = readSchemaPolicyInherits(
-      encodeSchema({
-        messages: {
-          columns: [{ name: "body", column_type: { type: "Text" }, nullable: false }],
-        },
-        reactions: {
-          columns: [
-            {
-              name: "message_id",
-              column_type: { type: "Uuid" },
-              nullable: false,
-              references: "messages",
-            },
-          ],
-          policies: {
-            delete: {
-              using: {
-                type: "Inherits",
-                operation: "Delete",
-                via_column: "message_id",
-              },
-            },
-          },
-        },
-      }),
-      "reactions",
-      "delete",
-    );
-
-    expect(policy).toEqual({
-      inherits: [{ parentColumn: "message_id" }],
-      joinCount: 0,
-    });
-  });
 });
 
 describe("NativeRuntimeAdapter TS adapter perf canary", () => {
@@ -5475,562 +4502,6 @@ function readPreparedFirstLiteral(query: Uint8Array): {
   const value = literalTag === 13 || literalTag === 14 ? Number(reader.i64()) : reader.u64();
   return { column, opTag, literalTag, value };
 }
-
-function readSchemaSelectPolicyBranches(
-  schemaBytes: Uint8Array,
-  tableName: string,
-): {
-  table: string;
-  filters: TestPolicyPredicate[];
-  joins: TestPolicyJoin[];
-  branches: TestPolicyBranch[];
-} {
-  return readSchemaPolicyBranches(schemaBytes, tableName, "select");
-}
-
-function readSchemaPolicyBranches(
-  schemaBytes: Uint8Array,
-  tableName: string,
-  operation: "select" | "insert" | "updateUsing" | "updateCheck" | "delete",
-): {
-  table: string;
-  filters: TestPolicyPredicate[];
-  joins: TestPolicyJoin[];
-  branches: TestPolicyBranch[];
-} {
-  const reader = new PostcardReader(schemaBytes);
-  const tables = reader.readVec((tableReader) => {
-    const table = tableReader.string();
-    tableReader.readVec((columnReader) => {
-      columnReader.string();
-      skipSchemaValueType(columnReader);
-      columnReader.option(() => undefined);
-      columnReader.option(() => undefined);
-      columnReader.option(skipGrooveValue);
-    });
-    const referenceCount = tableReader.u64();
-    for (let index = 0; index < referenceCount; index += 1) {
-      tableReader.string();
-      tableReader.string();
-    }
-    const policies = {
-      select: tableReader.option(readPolicyQueryForTest),
-      insert: tableReader.option(readPolicyQueryForTest),
-      updateUsing: tableReader.option(readPolicyQueryForTest),
-      updateCheck: tableReader.option(readPolicyQueryForTest),
-      delete: tableReader.option(readPolicyQueryForTest),
-    };
-    tableReader.u64();
-    const indexCount = tableReader.u64();
-    for (let index = 0; index < indexCount; index += 1) {
-      tableReader.string();
-      tableReader.readVec((indexReader) => indexReader.string());
-    }
-    return { table, policy: policies[operation] };
-  });
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-
-  const policy = tables.find((table) => table.table === tableName)?.policy;
-  expect(policy).toBeDefined();
-  return policy!;
-}
-
-function readSchemaColumnLargeValues(
-  schemaBytes: Uint8Array,
-  tableName: string,
-): Array<{ name: string; largeValue: "Text" | "Blob" | null }> {
-  const reader = new PostcardReader(schemaBytes);
-  const tables = reader.readVec((tableReader) => {
-    const table = tableReader.string();
-    const columns = tableReader.readVec((columnReader) => {
-      const name = columnReader.string();
-      skipSchemaValueType(columnReader);
-      const largeValue =
-        columnReader.option((kindReader) => {
-          const tag = kindReader.u64();
-          if (tag === 0) return "Text" as const;
-          if (tag === 1) return "Blob" as const;
-          throw new Error(`unsupported large value kind ${tag}`);
-        }) ?? null;
-      columnReader.option(() => undefined);
-      columnReader.option(skipGrooveValue);
-      return { name, largeValue };
-    });
-    const referenceCount = tableReader.u64();
-    for (let index = 0; index < referenceCount; index += 1) {
-      tableReader.string();
-      tableReader.string();
-    }
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.u64();
-    const indexCount = tableReader.u64();
-    for (let index = 0; index < indexCount; index += 1) {
-      tableReader.string();
-      tableReader.readVec((indexReader) => indexReader.string());
-    }
-    return { table, columns };
-  });
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  return tables.find((table) => table.table === tableName)?.columns ?? [];
-}
-
-function readSchemaTableMetadata(
-  schemaBytes: Uint8Array,
-  tableName: string,
-): {
-  indexedColumns: string[];
-  mergeStrategies: Array<{ column: string; strategy: "Lww" | "Counter" }>;
-} {
-  const reader = new PostcardReader(schemaBytes);
-  const tables = reader.readVec((tableReader) => {
-    const table = tableReader.string();
-    tableReader.readVec((columnReader) => {
-      columnReader.string();
-      skipSchemaValueType(columnReader);
-      columnReader.option(() => undefined);
-      columnReader.option(() => undefined);
-      columnReader.option(skipGrooveValue);
-    });
-    const referenceCount = tableReader.u64();
-    for (let index = 0; index < referenceCount; index += 1) {
-      tableReader.string();
-      tableReader.string();
-    }
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    tableReader.option(readPolicyQueryForTest);
-    const indexedColumns = tableReader.readVec((indexReader) => indexReader.string());
-    const mergeStrategyCount = tableReader.u64();
-    const mergeStrategies: Array<{ column: string; strategy: "Lww" | "Counter" }> = [];
-    for (let index = 0; index < mergeStrategyCount; index += 1) {
-      const column = tableReader.string();
-      const tag = tableReader.u64();
-      const strategy = tag === 0 ? "Lww" : tag === 1 ? "Counter" : null;
-      if (strategy == null) {
-        throw new Error(`unsupported merge strategy tag ${tag}`);
-      }
-      mergeStrategies.push({ column, strategy });
-    }
-    return { table, indexedColumns, mergeStrategies };
-  });
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  const table = tables.find((entry) => entry.table === tableName);
-  expect(table).toBeDefined();
-  return {
-    indexedColumns: table!.indexedColumns,
-    mergeStrategies: table!.mergeStrategies,
-  };
-}
-
-function readPolicyQueryForTest(reader: PostcardReader): {
-  table: string;
-  filters: TestPolicyPredicate[];
-  joins: TestPolicyJoin[];
-  branches: TestPolicyBranch[];
-} {
-  const table = reader.string();
-  const filters = reader.readVec(readPolicyPredicateForTest);
-  const joins = reader.readVec(readPolicyJoinForTest);
-  reader.option(() => undefined);
-  const branches = reader.readVec(readPolicyBranchForTest);
-  reader.readVec(skipPolicyReachableForTest);
-  reader.readVec(readPolicyInheritsForTest);
-  reader.readVec(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  reader.u64();
-  return { table, filters, joins, branches };
-}
-
-function readSchemaSelectPolicyReachables(
-  schemaBytes: Uint8Array,
-  tableName: string,
-): TestPolicyReachable[] {
-  const reader = new PostcardReader(schemaBytes);
-  const tables = reader.readVec((tableReader) => {
-    const table = tableReader.string();
-    tableReader.readVec((columnReader) => {
-      columnReader.string();
-      skipSchemaValueType(columnReader);
-      columnReader.option(() => undefined);
-      columnReader.option(() => undefined);
-      columnReader.option(skipGrooveValue);
-    });
-    const referenceCount = tableReader.u64();
-    for (let index = 0; index < referenceCount; index += 1) {
-      tableReader.string();
-      tableReader.string();
-    }
-    const select = tableReader.option(readPolicyQueryWithReachablesForTest);
-    tableReader.option(readPolicyQueryWithReachablesForTest);
-    tableReader.option(readPolicyQueryWithReachablesForTest);
-    tableReader.option(readPolicyQueryWithReachablesForTest);
-    tableReader.option(readPolicyQueryWithReachablesForTest);
-    tableReader.u64();
-    const indexCount = tableReader.u64();
-    for (let index = 0; index < indexCount; index += 1) {
-      tableReader.string();
-      tableReader.readVec((indexReader) => indexReader.string());
-    }
-    return { table, select };
-  });
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  return tables.find((table) => table.table === tableName)?.select?.reachables ?? [];
-}
-
-function readSchemaSelectPolicyInherits(
-  schemaBytes: Uint8Array,
-  tableName: string,
-): { inherits: TestPolicyInherits[]; joinCount: number } {
-  return readSchemaPolicyInherits(schemaBytes, tableName, "select");
-}
-
-function readSchemaPolicyInherits(
-  schemaBytes: Uint8Array,
-  tableName: string,
-  operation: "select" | "insert" | "updateUsing" | "updateCheck" | "delete",
-): { inherits: TestPolicyInherits[]; joinCount: number } {
-  const reader = new PostcardReader(schemaBytes);
-  const tables = reader.readVec((tableReader) => {
-    const table = tableReader.string();
-    tableReader.readVec((columnReader) => {
-      columnReader.string();
-      skipSchemaValueType(columnReader);
-      columnReader.option(() => undefined);
-      columnReader.option(() => undefined);
-      columnReader.option(skipGrooveValue);
-    });
-    const referenceCount = tableReader.u64();
-    for (let index = 0; index < referenceCount; index += 1) {
-      tableReader.string();
-      tableReader.string();
-    }
-    const policies = {
-      select: tableReader.option(readPolicyQueryWithInheritsForTest),
-      insert: tableReader.option(readPolicyQueryWithInheritsForTest),
-      updateUsing: tableReader.option(readPolicyQueryWithInheritsForTest),
-      updateCheck: tableReader.option(readPolicyQueryWithInheritsForTest),
-      delete: tableReader.option(readPolicyQueryWithInheritsForTest),
-    };
-    tableReader.u64();
-    const indexCount = tableReader.u64();
-    for (let index = 0; index < indexCount; index += 1) {
-      tableReader.string();
-      tableReader.readVec((indexReader) => indexReader.string());
-    }
-    return { table, policy: policies[operation] };
-  });
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  return (
-    tables.find((table) => table.table === tableName)?.policy ?? { inherits: [], joinCount: 0 }
-  );
-}
-
-function readPolicyQueryWithReachablesForTest(reader: PostcardReader): {
-  reachables: TestPolicyReachable[];
-} {
-  reader.string();
-  reader.readVec(readPolicyPredicateForTest);
-  reader.readVec(readPolicyJoinForTest);
-  reader.option(() => undefined);
-  reader.readVec(readPolicyBranchForTest);
-  const reachables = reader.readVec(readPolicyReachableForTest);
-  reader.readVec(() => undefined);
-  reader.readVec(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  reader.u64();
-  return { reachables };
-}
-
-function readPolicyQueryWithInheritsForTest(reader: PostcardReader): {
-  inherits: TestPolicyInherits[];
-  joinCount: number;
-} {
-  reader.string();
-  reader.readVec(readPolicyPredicateForTest);
-  const joinCount = reader.readVec(readPolicyJoinForTest).length;
-  reader.option(() => undefined);
-  reader.readVec(readPolicyBranchForTest);
-  reader.readVec(skipPolicyReachableForTest);
-  const inherits = reader.readVec(readPolicyInheritsForTest);
-  reader.readVec(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.readVec(() => undefined);
-  reader.option(() => undefined);
-  reader.option(() => undefined);
-  reader.u64();
-  return { inherits, joinCount };
-}
-
-function readPolicyBranchForTest(reader: PostcardReader): TestPolicyBranch {
-  const filters = reader.readVec(readPolicyPredicateForTest);
-  const joins = reader.readVec(readPolicyJoinForTest);
-  reader.readVec(skipPolicyReachableForTest);
-  reader.readVec(readPolicyInheritsForTest);
-  return { filters, joins };
-}
-
-function readPolicyInheritsForTest(reader: PostcardReader): TestPolicyInherits {
-  const parentColumn = reader.string();
-  reader.u64();
-  const maxDepth = reader.option((optionReader) => optionReader.u64());
-  return maxDepth === undefined ? { parentColumn } : { parentColumn, maxDepth };
-}
-
-function readPolicyJoinForTest(reader: PostcardReader): TestPolicyJoin {
-  const table = reader.string();
-  const onColumn = reader.string();
-  const targetTag = reader.u64();
-  const sourceColumn = reader.option((sourceReader) => sourceReader.string());
-  const sourceLookup = reader.option((lookupReader) => ({
-    table: lookupReader.string(),
-    rowIdSourceColumn: lookupReader.string(),
-    valueColumn: lookupReader.string(),
-  }));
-  reader.readVec((correlationReader) => {
-    correlationReader.string();
-    correlationReader.string();
-  });
-  const filters = reader.readVec(readPolicyPredicateForTest);
-  const nestedJoins = reader.readVec(readPolicyJoinForTest);
-  return { table, onColumn, targetTag, sourceColumn, sourceLookup, filters, nestedJoins };
-}
-
-function skipPolicyReachableForTest(reader: PostcardReader): void {
-  readPolicyReachableForTest(reader);
-}
-
-function readPolicyReachableForTest(reader: PostcardReader): TestPolicyReachable {
-  const accessTable = reader.string();
-  const accessRowColumn = reader.string();
-  const accessTeamColumn = reader.string();
-  const accessTeamTargetTag = reader.u64();
-  readPolicyOperandForTest(reader);
-  const accessFilters = reader.readVec(readPolicyPredicateForTest);
-  const edgeTable = reader.string();
-  const edgeMemberColumn = reader.string();
-  const edgeParentColumn = reader.string();
-  const edgeFilters = reader.readVec(readPolicyPredicateForTest);
-  const boundTag = reader.u64();
-  const maxDepth = boundTag === 1 ? reader.u64() : 0;
-  const seed = reader.option((seedReader) => ({
-    table: seedReader.string(),
-    userColumn: seedReader.option((userColumnReader) => userColumnReader.string()),
-    userClaim: seedReader.option((userClaimReader) => userClaimReader.string()),
-    teamColumn: seedReader.string(),
-    filters: seedReader.readVec(readPolicyPredicateForTest),
-  }));
-  return {
-    accessTable,
-    accessRowColumn,
-    accessTeamColumn,
-    accessTeamTargetTag,
-    accessFilters,
-    edgeTable,
-    edgeMemberColumn,
-    edgeParentColumn,
-    edgeFilters,
-    maxDepth,
-    seed,
-  };
-}
-
-function readPolicyPredicateForTest(reader: PostcardReader): TestPolicyPredicate {
-  const tag = reader.u64();
-  if (tag === 0 || tag === 1) {
-    return { tag, children: reader.readVec(readPolicyPredicateForTest) };
-  }
-  if (tag === 2) {
-    return { tag, child: readPolicyPredicateForTest(reader) };
-  }
-  if ([3, 4, 6, 7, 8, 9].includes(tag)) {
-    expect(reader.u64()).toBe(0);
-    const column = reader.string();
-    return { tag, column, operand: readPolicyOperandForTest(reader) };
-  }
-  if (tag === 5) {
-    readPolicyOperandForTest(reader);
-    reader.readVec(readPolicyOperandForTest);
-    return { tag };
-  }
-  if (tag === 10) {
-    readPolicyOperandForTest(reader);
-    readPolicyOperandForTest(reader);
-    return { tag };
-  }
-  if (tag === 11) {
-    readPolicyOperandForTest(reader);
-    return { tag };
-  }
-  throw new Error(`unsupported policy predicate tag ${tag}`);
-}
-
-function readPolicyOperandForTest(reader: PostcardReader): TestPolicyOperand {
-  const tag = reader.u64();
-  if (tag === 0) return { tag, column: reader.string() };
-  if (tag === 2) return { tag, claim: reader.string() };
-  if (tag === 3) {
-    const literalTag = reader.u64();
-    if (literalTag === 2 || literalTag === 3) {
-      return { tag, literalTag, value: reader.u64() };
-    }
-    if (literalTag === 13 || literalTag === 14) {
-      return { tag, literalTag, value: reader.i64() };
-    }
-    if (literalTag === 4) {
-      return { tag, literalTag, value: reader.bytes() };
-    }
-    if (literalTag === 5) {
-      return { tag, literalTag, value: reader.bool() };
-    }
-    if (literalTag === 6) {
-      return { tag, literalTag, value: reader.string() };
-    }
-    if (literalTag === 8) {
-      return { tag, literalTag, value: reader.bytes() };
-    }
-    if (literalTag === 12) {
-      return { tag, literalTag, value: reader.option(readPolicyOperandForTest) };
-    }
-    throw new Error(`unsupported policy literal tag ${literalTag}`);
-  }
-  throw new Error(`unsupported policy operand tag ${tag}`);
-}
-
-function skipSchemaValueType(reader: PostcardReader): void {
-  const tag = reader.u64();
-  if (tag === 11) {
-    reader.string();
-    reader.readVec((variant) => variant.string());
-    return;
-  }
-  if (tag === 12) {
-    reader.readVec(skipSchemaValueType);
-    return;
-  }
-  if (tag === 13 || tag === 14) {
-    skipSchemaValueType(reader);
-    return;
-  }
-  if (tag === 15) {
-    reader.readVec((field) => {
-      field.option((name) => name.string());
-      skipSchemaValueType(field);
-    });
-  }
-}
-
-function skipGrooveValue(reader: PostcardReader): void {
-  const tag = reader.u64();
-  switch (tag) {
-    case 0:
-    case 1:
-    case 2:
-    case 3:
-    case 9:
-      reader.u64();
-      return;
-    case 4:
-      reader.f64Le();
-      return;
-    case 5:
-      reader.bool();
-      return;
-    case 6:
-      reader.string();
-      return;
-    case 7:
-      reader.bytes();
-      return;
-    case 8:
-      reader.bytes(false);
-      return;
-    case 10:
-    case 11:
-      reader.readVec(skipGrooveValue);
-      return;
-    case 12:
-      reader.option(skipGrooveValue);
-      return;
-    case 13:
-      reader.i64();
-      return;
-    default:
-      throw new Error(`unsupported groove value tag ${tag}`);
-  }
-}
-
-type TestPolicyBranch = {
-  filters: TestPolicyPredicate[];
-  joins: TestPolicyJoin[];
-};
-
-type TestPolicyInherits = {
-  parentColumn: string;
-  maxDepth?: number;
-};
-
-type TestPolicyJoin = {
-  table: string;
-  onColumn: string;
-  targetTag: number;
-  sourceColumn: string | undefined;
-  sourceLookup: { table: string; rowIdSourceColumn: string; valueColumn: string } | undefined;
-  filters: TestPolicyPredicate[];
-  nestedJoins: TestPolicyJoin[];
-};
-
-type TestPolicyReachable = {
-  accessTable: string;
-  accessRowColumn: string;
-  accessTeamColumn: string;
-  accessTeamTargetTag: number;
-  accessFilters: TestPolicyPredicate[];
-  edgeTable: string;
-  edgeMemberColumn: string;
-  edgeParentColumn: string;
-  edgeFilters: TestPolicyPredicate[];
-  maxDepth: number;
-  seed:
-    | {
-        table: string;
-        userColumn: string | undefined;
-        userClaim: string | undefined;
-        teamColumn: string;
-        filters: TestPolicyPredicate[];
-      }
-    | undefined;
-};
-
-type TestPolicyPredicate =
-  | { tag: number; children: TestPolicyPredicate[] }
-  | { tag: number; child: TestPolicyPredicate }
-  | { tag: number; column: string; operand: TestPolicyOperand }
-  | { tag: number };
-
-type TestPolicyOperand =
-  | { tag: number; column: string }
-  | { tag: number; claim: string }
-  | { tag: number; literalTag: number; value: unknown };
 
 function unsupportedJoinRelationIr(): unknown {
   return {
